move mapspec editor
